+++ .idea/misc.xml
@@ -7,4 +7,11 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
+  <component name="SuppressABINotification">
+    <option name="modulesWithSuppressedNotConfigured">
+      <set>
+        <option value="ArcAssessmentsAndroid.sage" />
+      </set>
+    </option>
+  </component>
 </project>
