+++ SampleApp/build.gradle
@@ -30,6 +30,7 @@ android {
 dependencies {
     // Local projects
     implementation project(':arc')
+    implementation project(':sage')
     //implementation 'edu.wustl.arc.assessments:core-library:1.0.0'
     implementation 'com.github.gcacace:signature-pad:1.2.1'
     implementation 'com.google.code.gson:gson:2.8.9'
