+++ sage/build.gradle
@@ -1,55 +1,125 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
+apply plugin: 'com.android.library'
+apply plugin: 'maven-publish'
+apply plugin: 'kotlin-android'
+
+def versionMajor = 1 // max=2000 // bump to mark substantial changes
+def versionMinor = 0 // max=99   // bump for small changes (features, changes in intended behavior, etc)
+def versionPatch = 0 // max=99   // bump for fixes (crashes, unintended behavior, etc)
+def versionBuild = 0 // max=99   // bump for qa builds
+
+def flavor = getGradle().getStartParameter().getTaskRequests().toString().toLowerCase()
+
+def getVersionName = { ->
+    return "1.0.0"
+}
+
+def getArtificatId = { ->
+    return "sage-library" // Replace with library name ID
 }
 
 android {
-    namespace 'edu.wustl.arc.assessments.sage'
-    compileSdk 33
+
+    namespace = "edu.wustl.arc.sageassessments"
+    compileSdkVersion 33
+    flavorDimensions "default"
 
     defaultConfig {
-        minSdk 23
-        targetSdk 33
+        minSdkVersion 23
+        targetSdkVersion 33
+        versionCode versionMajor * 1000000 + versionMinor * 10000 + versionPatch * 100 + versionBuild
+        versionName "${getVersionName}"
+        buildConfigField "int", "ARC_VERSION_CODE", versionCode+""
+        buildConfigField "String", "ARC_VERSION_NAME", "\""+versionName+"\""
+        testInstrumentationRunner "android.support.session.runner.AndroidJUnitRunner"
+    }
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
+    compileOptions {
+        sourceCompatibility = 1.8
+        targetCompatibility = 1.8
+    }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
     }
 
     buildTypes {
+        debug {
+
+        }
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
+
+    lintOptions {
+        abortOnError false
     }
 }
 
 dependencies {
-    // WashU Arc
-    implementation("edu.wustl.arc.assessments:core-library:1.0.0")
-    implementation("com.github.gcacace:signature-pad:1.3.1")
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
 
-    // Sage
+    implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.4.0")
     implementation("org.sagebionetworks.assessmentmodel:presentation:0.10.1")
     implementation("org.sagebionetworks.assessmentmodel:assessmentModel:0.10.1")
+
+    implementation("edu.wustl.arc.assessments:core-library:1.0.0")
+    implementation 'com.github.gcacace:signature-pad:1.3.1'
+
+//    testImplementation project(':test_data')
+    testImplementation 'junit:junit:4.13'
+    testImplementation 'joda-time:joda-time:2.9.9'
+
+    api 'net.danlew:android.joda:2.9.9'
+    api 'com.squareup.retrofit2:converter-gson:2.5.0'
+    api 'com.squareup.retrofit2:retrofit:2.5.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    api 'androidx.appcompat:appcompat:1.1.0'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
+    api 'androidx.mediarouter:mediarouter:1.1.0'
+    api 'com.google.android.material:material:1.3.0'
+
+    implementation("androidx.compose.ui:ui:1.2.0-rc02")
+    implementation("androidx.compose.material:material:1.2.0-rc02")
+    implementation("androidx.compose.ui:ui-tooling-preview:1.2.0-rc02")
+    implementation("androidx.core:core-ktx:1.5.0")
 
     // Koin
     implementation("io.insert-koin:koin-core:3.1.5")
     implementation("io.insert-koin:koin-android:3.1.5")
 
-    // Default
-    implementation 'androidx.core:core-ktx:1.9.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.8.0'
-    implementation 'androidx.compose.ui:ui:1.3.3'
-    implementation 'com.google.code.gson:gson:2.8.9'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    // Kotlin
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.0")
 }
+
+publishing {
+    publications {
+        release(MavenPublication) {
+            groupId 'edu.wustl.arc.sageassessments' // Replace with group ID
+            artifactId getArtificatId()
+            version getVersionName()
+            artifact("$buildDir/outputs/aar/sage-release.aar")
+        }
+    }
+
+    repositories {
+        maven {
+            name = "GitHubPackages"
+            /** Configure path of your package repository on Github
+             ** Replace GITHUB_USERID with your/organisation Github userID
+             ** and REPOSITORY with the repository name on GitHub
+             */
+            url = uri("https://maven.pkg.github.com/CTRLab-WashU/ArcAssessmentsAndroid")
+            credentials {
+                /** Create github.properties in root project folder file with
+                 ** gpr.usr=GITHUB_USER_ID & gpr.key=PERSONAL_ACCESS_TOKEN
+                 ** Set env variable GPR_USER & GPR_API_KEY if not adding a properties file**/
+
+                username = project.findProperty('gpr.usr') ?: System.getenv("GPR_USER")
+                password = project.findProperty('gpr.key') ?: System.getenv("GPR_TOKEN")
+            }
+        }
+    }
+}
+
