+++ sage/src/main/AndroidManifest.xml
@@ -1,4 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    package="edu.wustl.arc.sageassessments">
+
+    <application
+        android:allowBackup="false" >
+    </application>
 
 </manifest>
